UNST-5035 unstrucversion = 1.2.125    ! date  = 20210604                             ! UNST-5035, removelinksofhangingnodes, connecthangingnodes

+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_utils/rest_f90/readarcinfoblock.f90
@@ -39,7 +39,7 @@
       integer :: minp
       integer :: nc
       double precision :: rmis
-      double precision :: D(MC,NC)
+      real    :: D(MC,NC)
       double precision, dimension(MC) :: dline
       CHARACTER TEX*16
